client/requirements.txt
bidict==0.21.4
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-engineio==4.3.1
python-socketio==5.5.1
pyxel==1.6.1
requests==2.27.1
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
websocket-client==1.2.3
